% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sqlscore.R
\name{sqlscore}
\alias{sqlscore}
\title{Convert an R object into a list of SQL queries}
\usage{
sqlscore(x, ...)
}
\arguments{
\item{x}{An R object to turn into SQL queries}

\item{...}{Arguments passed to the scoring method}
}
\description{
S3 generic for turning R obkect into SQL text
}

