% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sqlscore.train.R
\name{sqlscore.train}
\alias{sqlscore.train}
\title{Convert an train object into a list of SQL queries}
\usage{
\method{sqlscore}{train}(x, ...)
}
\arguments{
\item{x}{A caret object}

\item{...}{Arguments passed to the scoring method}
}
\description{
Runs the correct sqlscore method on x$finalModel
}

